serial_buffer: Move the socket console out of the serial manager
Move the socket console out of the serial manager into serial_buffer so
it can be reused by other implementations, such as the virtio console.
